+++ mes-java/mes-domain/src/main/java/com/zeroone/star/project/j5/query/shiftplan/ShiftPlanQuery.java
@@ -13,5 +13,5 @@ public class ShiftPlanQuery extends PageQuery {
      * 排班计划id
      */
     @ApiModelProperty(value = "排班计划id", example = "1")
-    private Integer PlanId;
+    private Long PlanId;
 }
